Minify map filter panel
